Move default quantizer
