+++ cluster/schema/manager.go
@@ -169,7 +169,7 @@ func (s *SchemaManager) Close(ctx context.Context) (err error) {
 	return s.db.Close(ctx)
 }
 
-func (s *SchemaManager) AddClass(cmd *command.ApplyRequest, nodeID string, schemaOnly bool, enableSchemaCallback bool, defaultQuantization *configRuntime.DynamicValue[int]) error {
+func (s *SchemaManager) AddClass(cmd *command.ApplyRequest, nodeID string, schemaOnly bool, enableSchemaCallback bool) error {
 	req := command.AddClassRequest{}
 	// dupa
 	if err := json.Unmarshal(cmd.SubCommand, &req); err != nil {
@@ -187,9 +187,6 @@ func (s *SchemaManager) AddClass(cmd *command.ApplyRequest, nodeID string, schem
 	// tenant) we don't want another goroutine holding a pointer to it and finding issues with concurrent read/writes.
 	shardingStateCopy := req.State.DeepCopy()
 
-	// Enable default quantization for each vector index config if no quantization is set by the user
-	s.enableQuantization(req.Class, defaultQuantization)
-
 	return s.apply(
 		applyOp{
 			op:                   cmd.GetType().String(),
@@ -203,35 +200,6 @@ func (s *SchemaManager) AddClass(cmd *command.ApplyRequest, nodeID string, schem
 
 func hasTargetVectors(class *models.Class) bool {
 	return len(class.VectorConfig) > 0
-}
-
-func (s *SchemaManager) enableQuantization(class *models.Class, defaultQuantization *configRuntime.DynamicValue[int]) {
-	compression := dimensioncategory.DimensionCategory(defaultQuantization.Get()).String()
-	if !hasTargetVectors(class) || class.VectorIndexType != "" {
-		setDefaultQuantization(class.VectorIndexConfig, compression)
-	}
-
-	for _, vectorConfig := range class.VectorConfig {
-		setDefaultQuantization(vectorConfig.VectorIndexConfig, compression)
-	}
-
-}
-
-func setDefaultQuantization(vectorIndexConfig interface{}, compression string) {
-	found := false
-	for _, quantization := range []string{"pq", "sq", "rq", "bq"} {
-		quantizationConfig := vectorIndexConfig.(map[string]interface{})[quantization]
-		if _, ok := quantizationConfig.(map[string]interface{})["enabled"]; ok {
-			found = true
-			break
-		}
-	}
-	if !found {
-		vectorIndexConfig.(map[string]interface{})[compression] = map[string]interface{}{
-			"enabled": true,
-		}
-		vectorIndexConfig.(map[string]interface{})["trackingDefault"] = true
-	}
 }
 
 func (s *SchemaManager) RestoreClass(cmd *command.ApplyRequest, nodeID string, schemaOnly bool, enableSchemaCallback bool) error {
